+++ R/src/library/tcltk/src/tcltk_unix.c
@@ -141,8 +141,7 @@ void Tcl_unix_setup(void)
 
 /* Fill a text buffer with user typed console input. */
 
-static int
-RTcl_ReadConsole (const char *prompt, unsigned char *buf, int len,
+static int RTcl_ReadConsole (const char *prompt, unsigned char *buf, size_t len,
 		  int addtohistory)
 {
     Tcl_Obj *cmd[3];
